Extract internal utilities to separate library
