+++ ayza/pom.xml
@@ -48,6 +48,10 @@
             <groupId>io.github.hakky54</groupId>
             <artifactId>sude</artifactId>
         </dependency>
+        <dependency>
+            <groupId>io.github.hakky54</groupId>
+            <artifactId>laleler</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.junit.jupiter</groupId>
