--- ayza/src/main/java/nl/altindag/ssl/util/internal/StringUtils.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2019 Thunderberry.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package nl.altindag.ssl.util.internal;
-
-import static java.util.Objects.isNull;
-
-/**
- * <strong>NOTE:</strong>
- * Please don't use this class directly as it is part of the internal API. Class name and methods can be changed any time.
- *
- * @author Hakan Altindag
- */
-public final class StringUtils {
-
-    private StringUtils() {}
-
-    public static boolean isBlank(CharSequence charSequence) {
-        int length = isNull(charSequence) ? 0 : charSequence.length();
-        if (length != 0) {
-            for (int i = 0; i < length; ++i) {
-                if (!Character.isWhitespace(charSequence.charAt(i))) {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-    public static boolean isNotBlank(CharSequence charSequence) {
-        return !isBlank(charSequence);
-    }
-
-}
