+++ pom.xml
@@ -55,6 +55,7 @@
 
     <properties>
         <version.sude>2.0.2</version.sude>
+        <version.laleler>1.0.0</version.laleler>
         <version.slf4j>2.0.17</version.slf4j>
         <version.junit>5.14.4</version.junit>
         <version.assertj-core>3.27.7</version.assertj-core>
@@ -80,6 +81,11 @@
                 <groupId>io.github.hakky54</groupId>
                 <artifactId>sude</artifactId>
                 <version>${version.sude}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.github.hakky54</groupId>
+                <artifactId>laleler</artifactId>
+                <version>${version.laleler}</version>
             </dependency>
 
             <dependency>
